# sql: Make JSON parsing faster.