--- a/pkg/util/json/BUILD.bazel
+++ b/pkg/util/json/BUILD.bazel
@@ -5,6 +5,7 @@ load("//pkg/testutils/buildutil:buildutil.bzl", "disallowed_imports_test")
 go_library(
     name = "json",
     srcs = [
+        "config.go",
         "contains.go",
         "contains_testers.go",
         "encode.go",
@@ -12,6 +13,7 @@ go_library(
         "iterator.go",
         "jentry.go",
         "json.go",
+        "parser.go",
         "random.go",
         "tables.go",
     ],
@@ -26,6 +28,8 @@ go_library(
         "//pkg/sql/pgwire/pgerror",
         "//pkg/util",
         "//pkg/util/encoding",
+        "//pkg/util/json/tokenizer",
+        "//pkg/util/randutil",
         "//pkg/util/syncutil",
         "//pkg/util/unique",
         "@com_github_cockroachdb_apd_v3//:apd",