--- a/pkg/util/json/tokenizer/BUILD.bazel
+++ b/pkg/util/json/tokenizer/BUILD.bazel
@@ -1,0 +1,28 @@
+load("//build/bazelutil/unused_checker:unused.bzl", "get_x_data")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "tokenizer",
+    srcs = [
+        "decoder.go",
+        "reader.go",
+        "scanner.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/util/json/tokenizer",
+    visibility = ["//visibility:public"],
+)
+
+go_test(
+    name = "tokenizer_test",
+    srcs = [
+        "bench_test.go",
+        "decoder_test.go",
+        "reader_test.go",
+        "scanner_test.go",
+    ],
+    args = ["-test.timeout=295s"],
+    data = glob(["testdata/**"]),
+    embed = [":tokenizer"],
+)
+
+get_x_data(name = "get_x_data")