`tree.ParseJSON` is slow, and inefficient, primary due to its multi step
parsing implementation, where first, the input JSON string is decoded using 
Go standard `encoding/json` library, and then the Go native data type is 
converted to `tree.JSON`.

There are many inefficiencies in this process including reliance on reflection, 
unnecessary allocations (map[string]interface{}) and many more.  
The reality is that `encoding/json` does too many things that we do not need 
(i.e ability to decode into `struct{}`) and that hurts performance.

This PR introduces a new fast json parser implementation,
to directly convert input string into the `tree.JSON` representation. 
There are no intermediate steps.  As we process input string, 
we build the final `tree.JSON` representation.
This implementation does not use reflection, and, in general,
is substantially faster, and more memory efficient across all benchmarks.

JSON parser is built on top of somewhat heavily modified fork of `pkg/json` package.  
That package, while incomplete, provided an excellent starting point.
Namely its `Scanner` and `Decoder` implementations provided a blueprint 
on how to implement the higher level logic around direct conversion into `tree.JSON`.
The techniques demonstrated in this package
(https://dave.cheney.net/high-performance-json.html) were used throughout 
this package in order to build high performance state machine to convert 
strings into JSONb. This package was modified, somewhat heavily, to include 
required, but missing parts, including string un-escaping, as well as handling of UTF characters.

New implementation should be a drop in replacement to our use of `encoder/json` in `tree.ParseJSON`.
Existing tests were modified to test both implementation.  Extensive improvement were 
made to random JSON generation library so that the tests are more complete and thorough.

Performance of this new library substantially exceeds Go standard implementation
for common cases -- namely, the case when input strings do not contain escape sequences. 
The new implementation consistently requires fewer (and smaller) allocations, 
while at the same time being faster both in terms of time per op, as well as raw throughput. 
The benchmark comparing large input, with strings fields up to 1/2MB in size, 
the difference in throughput was over 8 *times*.

Handling escaped, or UTF encoded data in JSON stream, is an expensive proposition: 
as soon as the data is known to contain escape characters `\`, additional memory allocations 
and copies are needed.  This hurts performance.  While the new implementation is still faster, 
and more memory efficient, the differences are not as pronounced (only 75% faster when
encoding heavily escaped, almost binary data).

The final benefit of the new implementation is that we may choose (in the future) to no longer
sort object keys upon construction.  This capability is exposed via a
`jsonb.WithUnorderedObjectKeys()` option to `tree.ParseJSON`. 
The standard Go implementation had to sort the keys because otherwise, 
the same input will not produce round trip-able output because of non-deterministic 
order of map traversal in Go.  This is no longer applicable to this implementation 
since the objects are constructed in a well defined ordered. This option, by itself, results
in a modest 5-7% improvement in speed (and about 10% reduction in allocations).
While these improvements are not that significant, more possibilities exists to push the
envelope, perhaps by supporting some custom parsing options.

Fixes #89998 

Release note (performance improvement): significantly faster JSON parsing.  
This should help with any workloads that insert large amounts of JSON data, including IMPORT workloads.