@miretskiy 2022-10-13
GoStandard vs Lexer implementation benchmarks are below.
The percentage part in the benchmark name indicates the amount of data
that required escaping: with 0% being the happy case, 3% modeling the case
when the data has modest amount of escaped data (few `"`, few utf characters, etc),
and a 40% simulating binary (or UTF8 heavy data).
The size portion indicates the maximum size of the strings inside generated
(random) JSON inputs.

While the speed impact is the most striking (particularly in the no-escape case),
the lexer implementation, across all benchmarks, has a much nicer and predictable
allocation behavior.

```
name                               old time/op    new time/op     delta
ParseJSON/escape=0.00%/128-32        5.50µs ± 5%     2.88µs ± 4%   -47.54%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/512-32        7.31µs ± 6%     3.19µs ± 1%   -56.39%  (p=0.000 n=10+8)
ParseJSON/escape=0.00%/4096-32       26.5µs ± 1%      7.2µs ± 2%   -72.76%  (p=0.000 n=8+9)
ParseJSON/escape=0.00%/16384-32      85.1µs ± 1%     18.3µs ± 1%   -78.49%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/65536-32       312µs ± 1%       62µs ± 2%   -80.12%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/262144-32     1.35ms ± 1%     0.23ms ± 0%   -82.98%  (p=0.000 n=8+9)
ParseJSON/escape=0.00%/524288-32     2.62ms ± 1%     0.43ms ± 3%   -83.66%  (p=0.000 n=9+9)
ParseJSON/escape=3.00%/128-32        5.83µs ± 3%     3.38µs ± 1%   -42.00%  (p=0.000 n=9+9)
ParseJSON/escape=3.00%/512-32        8.63µs ± 0%     5.58µs ± 0%   -35.33%  (p=0.000 n=7+8)
ParseJSON/escape=3.00%/4096-32       34.0µs ± 1%     24.9µs ± 1%   -26.72%  (p=0.000 n=9+8)
ParseJSON/escape=3.00%/16384-32       110µs ± 2%       88µs ± 1%   -20.53%  (p=0.000 n=9+10)
ParseJSON/escape=3.00%/65536-32       416µs ± 1%      342µs ± 0%   -17.87%  (p=0.000 n=9+9)
ParseJSON/escape=3.00%/262144-32     1.76ms ± 7%     1.38ms ± 1%   -21.86%  (p=0.000 n=8+9)
ParseJSON/escape=3.00%/524288-32     3.44ms ± 5%     2.79ms ± 1%   -18.79%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/128-32       6.52µs ± 3%     3.75µs ± 1%   -42.46%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/512-32       11.7µs ± 1%      6.7µs ± 1%   -42.54%  (p=0.000 n=8+8)
ParseJSON/escape=40.00%/4096-32      55.6µs ± 1%     32.1µs ± 0%   -42.31%  (p=0.000 n=10+8)
ParseJSON/escape=40.00%/16384-32      204µs ± 1%      117µs ± 1%   -42.43%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/65536-32      831µs ± 1%      469µs ± 0%   -43.53%  (p=0.000 n=10+8)
ParseJSON/escape=40.00%/262144-32    3.18ms ± 0%     1.83ms ± 1%   -42.28%  (p=0.000 n=7+10)
ParseJSON/escape=40.00%/524288-32    6.77ms ± 1%     3.60ms ± 0%   -46.78%  (p=0.000 n=10+10)

name                               old speed      new speed       delta
ParseJSON/escape=0.00%/128-32      30.6MB/s ± 4%   58.3MB/s ± 4%   +90.62%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/512-32      58.4MB/s ± 6%  133.7MB/s ± 1%  +129.04%  (p=0.000 n=10+8)
ParseJSON/escape=0.00%/4096-32      110MB/s ± 1%    405MB/s ± 2%  +267.14%  (p=0.000 n=8+9)
ParseJSON/escape=0.00%/16384-32     133MB/s ± 1%    618MB/s ± 1%  +364.94%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/65536-32     144MB/s ± 1%    723MB/s ± 2%  +402.76%  (p=0.000 n=9+10)
ParseJSON/escape=0.00%/262144-32    138MB/s ± 1%    810MB/s ± 0%  +488.64%  (p=0.000 n=8+9)
ParseJSON/escape=0.00%/524288-32    139MB/s ± 1%    862MB/s ± 3%  +520.11%  (p=0.000 n=9+9)
ParseJSON/escape=3.00%/128-32      29.5MB/s ± 3%   50.8MB/s ± 1%   +72.18%  (p=0.000 n=9+10)
ParseJSON/escape=3.00%/512-32      52.4MB/s ± 0%   81.0MB/s ± 0%   +54.64%  (p=0.000 n=7+8)
ParseJSON/escape=3.00%/4096-32     90.0MB/s ± 1%  122.9MB/s ± 1%   +36.48%  (p=0.000 n=9+8)
ParseJSON/escape=3.00%/16384-32     107MB/s ± 2%    135MB/s ± 1%   +25.74%  (p=0.000 n=9+10)
ParseJSON/escape=3.00%/65536-32     114MB/s ± 1%    139MB/s ± 0%   +21.74%  (p=0.000 n=9+9)
ParseJSON/escape=3.00%/262144-32    108MB/s ± 5%    139MB/s ± 1%   +28.12%  (p=0.000 n=7+9)
ParseJSON/escape=3.00%/524288-32    111MB/s ± 5%    138MB/s ± 1%   +24.48%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/128-32     29.9MB/s ± 3%   52.0MB/s ± 1%   +73.77%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/512-32     47.4MB/s ± 1%   82.5MB/s ± 1%   +74.05%  (p=0.000 n=8+8)
ParseJSON/escape=40.00%/4096-32    68.5MB/s ± 1%  118.8MB/s ± 0%   +73.33%  (p=0.000 n=10+8)
ParseJSON/escape=40.00%/16384-32   73.3MB/s ± 1%  127.3MB/s ± 1%   +73.68%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/65536-32   73.4MB/s ± 1%  130.3MB/s ± 0%   +77.40%  (p=0.000 n=10+8)
ParseJSON/escape=40.00%/262144-32  74.6MB/s ± 0%  129.6MB/s ± 0%   +73.63%  (p=0.000 n=7+9)
ParseJSON/escape=40.00%/524288-32  75.6MB/s ± 1%  132.6MB/s ± 0%   +75.37%  (p=0.000 n=10+10)

name                               old alloc/op   new alloc/op    delta
ParseJSON/escape=0.00%/128-32        2.44kB ± 0%     1.06kB ± 0%   -56.42%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/512-32        3.25kB ± 0%     1.32kB ± 0%   -59.26%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/4096-32       13.0kB ± 0%      4.0kB ± 0%   -69.08%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/16384-32      45.3kB ± 0%     12.8kB ± 0%   -71.71%  (p=0.000 n=10+8)
ParseJSON/escape=0.00%/65536-32       177kB ± 0%       49kB ± 0%   -72.17%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/262144-32      717kB ± 0%      191kB ± 0%   -73.35%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/524288-32     1.40MB ± 0%     0.37MB ± 0%   -73.31%  (p=0.000 n=9+10)
ParseJSON/escape=3.00%/128-32        2.56kB ± 0%     1.16kB ± 0%   -54.83%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/512-32        3.74kB ± 0%     1.71kB ± 0%   -54.19%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/4096-32       16.7kB ± 0%      7.1kB ± 0%   -57.71%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/16384-32      59.3kB ± 0%     24.6kB ± 0%   -58.47%  (p=0.000 n=9+10)
ParseJSON/escape=3.00%/65536-32       236kB ± 0%       99kB ± 0%   -58.11%  (p=0.000 n=8+10)
ParseJSON/escape=3.00%/262144-32      940kB ± 0%      376kB ± 0%   -60.02%  (p=0.000 n=8+10)
ParseJSON/escape=3.00%/524288-32     1.83MB ± 0%     0.75MB ± 0%   -59.16%  (p=0.000 n=10+9)
ParseJSON/escape=40.00%/128-32       2.62kB ± 0%     1.16kB ± 0%   -55.62%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/512-32       4.17kB ± 0%     1.74kB ± 0%   -58.28%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/4096-32      19.7kB ± 0%      7.0kB ± 0%   -64.32%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/16384-32     72.7kB ± 0%     24.9kB ± 0%   -65.80%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/65536-32      293kB ± 0%      101kB ± 0%   -65.55%  (p=0.000 n=9+9)
ParseJSON/escape=40.00%/262144-32    1.12MB ± 0%     0.37MB ± 1%   -66.76%  (p=0.000 n=9+10)
ParseJSON/escape=40.00%/524288-32    2.42MB ± 1%     0.74MB ± 0%   -69.48%  (p=0.000 n=10+10)

name                               old allocs/op  new allocs/op   delta
ParseJSON/escape=0.00%/128-32          39.0 ± 0%       29.0 ± 0%   -25.64%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/512-32          39.0 ± 0%       29.0 ± 0%   -25.64%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/4096-32         41.0 ± 0%       29.0 ± 0%   -29.27%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/16384-32        41.0 ± 0%       29.0 ± 0%   -29.27%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/65536-32        43.0 ± 0%       29.0 ± 0%   -32.56%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/262144-32       45.0 ± 0%       29.0 ± 0%   -35.56%  (p=0.000 n=10+10)
ParseJSON/escape=0.00%/524288-32       44.0 ± 0%       29.0 ± 0%   -34.09%  (p=0.002 n=8+10)
ParseJSON/escape=3.00%/128-32          40.0 ± 0%       30.0 ± 0%   -25.00%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/512-32          41.0 ± 0%       30.0 ± 0%   -26.83%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/4096-32         42.0 ± 0%       30.0 ± 0%   -28.57%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/16384-32        43.0 ± 0%       30.0 ± 0%   -30.23%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/65536-32        44.0 ± 0%       30.0 ± 0%   -31.82%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/262144-32       46.0 ± 0%       30.0 ± 0%   -34.78%  (p=0.000 n=10+10)
ParseJSON/escape=3.00%/524288-32       45.7 ± 2%       30.0 ± 0%   -34.35%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/128-32         40.0 ± 0%       30.0 ± 0%   -25.00%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/512-32         41.0 ± 0%       31.0 ± 0%   -24.39%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/4096-32        42.0 ± 0%       30.0 ± 0%   -28.57%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/16384-32       43.0 ± 0%       30.0 ± 0%   -30.23%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/65536-32       45.0 ± 0%       30.0 ± 0%   -33.33%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/262144-32      45.0 ± 0%       30.0 ± 0%   -33.33%  (p=0.000 n=10+10)
ParseJSON/escape=40.00%/524288-32      46.0 ± 0%       30.0 ± 0%   -34.78%  (p=0.000 n=9+10)
```